Extract and rename test
